requirements.txt
flask==3.0.3
werkzeug==3.0.3
Pillow==10.4.0
google-genai==1.5.0
numpy==1.26.4
opencv-contrib-python-headless==4.9.0.80
gunicorn==22.0.0
